fixed: store the gammacurve as floats instead of ints
git-svn-id: http://boblight.googlecode.com/svn/trunk@442 c8956a1e-8086-11de-bc80-f7d40095c7a4

+++ src/lib/boblight_client.h
@@ -52,7 +52,7 @@ namespace boblight
       float       m_satrange[2];
       int         m_threshold;
       float       m_gamma;
-      int         m_gammacurve[256];
+      float       m_gammacurve[256];
 
       float       m_rgb[3];
       int         m_rgbcount;
